requirements.txt
absl-py==0.11.0
astunparse==1.6.3
bcrypt==3.2.0
cachetools==4.1.1
click==7.1.2
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
gast==0.3.3
gunicorn==19.9.0
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.33.2
h5py==2.10.0
itsdangerous==1.1.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
Markdown==3.3.3
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
pandas==1.1.4
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2020.4
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.5.3
SQLAlchemy==1.3.20
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
Werkzeug==1.0.1